+++ opentreemap/treemap/templates/treemap/map.html
@@ -11,10 +11,10 @@
 <script>
 (function(require) {
     var $ = require("jquery"),
-        map = require("treemap/map");
+        treeMap = require("treemap/treeMap");
 
     $(function(){
-        map.initMapPage(otm.settings);
+        treeMap.init(otm.settings);
     })
 })(require);
 </script>
